Added a project that runs the engine will become a game

+++ Minigin/GameObject.h
@@ -39,7 +39,7 @@ void GameObject::AddComponent(component* pComp)
 template <class component>
 component* GameObject::GetComponent()
 {
-	for (int i = 0; i < m_Components.size(); i++ )
+	for (size_t i = 0; i < m_Components.size(); i++ )
 	{
 		component* pComponent = dynamic_cast<component*>(m_Components[i]);
 		if (pComponent != nullptr)
